@@ -1,20 +1,8 @@
 package by.it.medved.repositories;
 
 import by.it.medved.entities.Movie;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.List;
+public interface MovieRepository extends JpaRepository<Movie, Long> {
 
-public interface MovieRepository {
-
-    Movie addMovie(Movie movie);
-
-    Movie getMovieById(Long id);
-
-    List<Movie> getMovies();
-
-    Movie updateMovie(Long id, LocalDateTime showDateTime, BigDecimal price);
-
-    void deleteMovieById(Long id);
 }
