@@ -1,25 +1,13 @@
 package by.it.medved.repositories;
 
 import by.it.medved.entities.User;
-import by.it.medved.enums.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
-import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository {
+@Repository
+public interface UserRepository extends JpaRepository <User, Long> {
 
-    User createUser(User user);
-
-    User getUserById(Long id);
-
-    List<User> getUsers();
-
-    User changeUserPassword(Long userId, byte[] encryptedPassword);
-
-    boolean updateRole(Long id, Role role);
-
-    User updateUserFields(Long id, String firstName, String email, LocalDate dateBirthday);
-
-    void deleteUserById(Long id);
+    Optional<User> findUserByLogin(String login);
 }
