@@ -1,32 +1,22 @@
 package by.it.medved.services;
 
-import by.it.medved.entities.User;
+import by.it.medved.dto.UserRequest;
+import by.it.medved.dto.UserResponse;
 import by.it.medved.enums.Role;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface UserService {
 
-    User createUser(User user);
+    UserResponse saveUser(UserRequest userRequest);
 
-    User getUserById(Long id);
+    UserResponse getUserById(Long id);
 
-    List<User> getUsers();
+    UserResponse getUserByLogin(String login);
 
-    User changeUserPassword(String newPassword, Long userId);
+    List<UserResponse> getUsers();
 
-    boolean updateRole(Long id, Role role);
+    UserResponse updateRole(Long id, Role role);
 
-    User updateUserFields(Long id, String firstName, String email, String dateBirthday);
-
-    void deleteUserById(Long id);
-
-    Optional<User> getUserByLogin(String login);
-
-    boolean checkLine(String line, String regex);
-
-    boolean checkDate(String line, String regex);
-
-    boolean checkUniqueLogin(String login);
+    void deleteUser(Long id);
 }
